Extract chartconfig specific setup and teardown logic to chartconfig package
